+++ Common/src/main/java/com/hypherionmc/sdlink/core/discord/events/DiscordEventHandler.java
@@ -161,10 +161,7 @@ public final class DiscordEventHandler extends ListenerAdapter {
         try {
             List<SDLinkAccount> accounts = DatabaseManager.INSTANCE.getCollection(SDLinkAccount.class);
             Optional<SDLinkAccount> account = accounts.stream().filter(a -> a.getDiscordID() != null && a.getDiscordID().equalsIgnoreCase(event.getUser().getId())).findFirst();
-
-            account.ifPresent(a -> {
-                DatabaseManager.INSTANCE.deleteEntry(a, SDLinkAccount.class);
-            });
+            account.ifPresent(a -> DatabaseManager.INSTANCE.deleteEntry(a, SDLinkAccount.class));
         } catch (Exception e) {
             BotController.INSTANCE.getLogger().error("Failed to remove linked account", e);
         }